Updated convFlow blocks
